Refactor and prepare for the viewmodel
