+++ app/build.gradle
@@ -31,6 +31,7 @@ dependencies {
     implementation("androidx.recyclerview:recyclerview:1.1.0")
     implementation("com.google.android.material:material:1.0.0")
     implementation("com.google.android:flexbox:2.0.0")
+    implementation("android.arch.lifecycle:extensions:1.1.1")
     testImplementation("junit:junit:4.12")
     androidTestImplementation("androidx.test:runner:1.2.0")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.2.0")
